JWT extended updates

+++ resources/item.py
@@ -1,6 +1,6 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, get_jwt_identity, jwt_optional, fresh_jwt_required
 
 from models.item import ItemModel
 
@@ -15,7 +15,7 @@ class Item(Resource):
                         required=True,
                         help="Every item needs a store id")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         try:
             item = ItemModel.find_by_name(name)
@@ -26,6 +26,7 @@ class Item(Resource):
             return item.json()
         return {"message": "Item not found"}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400  # 400: bad request
@@ -40,13 +41,18 @@ class Item(Resource):
 
         return item.json(), 201  # http code for 'creating' = 201
 
-    def delete(self,name):
+    @jwt_required
+    def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
+
         item = ItemModel.find_by_name(name)
 
         if item:
            item.delete_from_db()
-
-        return {"message": "Item Deleted"}
+           return {"message": "Item Deleted"}
+        return {"message": "Item not found"}, 404
 
 
 
@@ -72,5 +78,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {"items": [x.json() for x in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [x.json() for x in ItemModel.find_all()]
+        if user_id:
+            return {"items": items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }
